+++ server/tests/main_test.rs
@@ -91,7 +91,7 @@ async fn run() {
     wait_until_data_loaded_after(zmq_endpoint, &start_test_datetime).await;
     wait_until_connected_to_rabbitmq(zmq_endpoint).await;
 
-    subtests::kirin_delete_vj_test::delete_vj_test(&config).await;
+    subtests::realtime_test::remove_add_modify_vj_test(&config).await;
 
     subtests::reload_test::reload_test(&config, &data_dir_path).await;
