+++ server/tests/subtests/mod.rs
@@ -27,6 +27,6 @@
 // https://groups.google.com/d/forum/navitia
 // www.navitia.io
 
-pub mod kirin_delete_vj_test;
 pub mod places_nearby_test;
+pub mod realtime_test;
 pub mod reload_test;
